backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
pydantic
requests
tiktoken

# Only install this on Linux (Render), skip on Windows
pysqlite3-binary ; sys_platform == 'linux'

# For GitHub and LinkedIn (PDF/Text) loading
pypdf
gitpython

